⚡️ Improve Login function
- 서버 주소를 상수로 분리했다
- 로그인시 엔터키를 통해 로그인 가능하도록 했다
