+++ packages/client/src/shared/lib/constants/api.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = 'https://www.별글.site/api/';
