+++ packages/client/src/shared/lib/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './camera';
 export * from './space';
+export * from './api';
