+++ packages/client/src/shared/ui/buttons/Button.tsx
@@ -4,7 +4,7 @@ import { css } from '@emotion/react';
 import { Body02ME, Body03ME, Body02BD, Body03BD } from '../styles';
 
 interface PropsType extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-	onClick: () => void;
+	onClick?: () => void;
 	children: React.ReactNode;
 	size: 'm' | 'l';
 	buttonType: 'Button' | 'CTA-icon' | 'warning' | 'warning-border';
