+++ packages/client/src/widgets/loginModal/ui/LeftButton.tsx
@@ -6,7 +6,7 @@ interface PropsType {
 
 export default function LeftButton({ onClick }: PropsType) {
 	return (
-		<Button onClick={onClick} size="m" buttonType="Button">
+		<Button onClick={onClick} size="m" buttonType="Button" type="button">
 			회원가입
 		</Button>
 	);
